move gohcl code into parse_hcl
