+++ pkg/steampipeconfig/parse/parse_hcl/decode.go
@@ -1,4 +1,4 @@
-package parse
+package parse_hcl
 
 import (
 	"fmt"
@@ -7,18 +7,7 @@ import (
 	"github.com/zclconf/go-cty/cty/convert"
 	"github.com/zclconf/go-cty/cty/gocty"
 	"reflect"
-	"sort"
-	"strings"
 )
-
-var victimExpr hcl.Expression
-var victimBody hcl.Body
-
-var exprType = reflect.TypeOf(&victimExpr).Elem()
-var bodyType = reflect.TypeOf(&victimBody).Elem()
-var blockType = reflect.TypeOf((*hcl.Block)(nil))
-var attrType = reflect.TypeOf((*hcl.Attribute)(nil))
-var attrsType = reflect.TypeOf(hcl.Attributes(nil))
 
 // DecodeBody extracts the configuration within the given body into the given
 // value. This value must be a non-nil pointer to either a struct or
@@ -57,6 +46,8 @@ func decodeBodyToValue(body hcl.Body, ctx *hcl.EvalContext, val reflect.Value) h
 	}
 }
 
+// NOTE: change in behaviour from gohcl - when evaluating an expression, call DecodeExpressionNested instead of
+// DecodeExpression - this will populate nested structs from the cty value
 func decodeBodyToStruct(body hcl.Body, ctx *hcl.EvalContext, val reflect.Value) hcl.Diagnostics {
 	schema, partial := ImpliedBodySchema(val.Interface())
 
@@ -130,7 +121,8 @@ func decodeBodyToStruct(body hcl.Body, ctx *hcl.EvalContext, val reflect.Value) 
 		case exprType.AssignableTo(field.Type):
 			fieldV.Set(reflect.ValueOf(attr.Expr))
 		default:
-			diags = append(diags, DecodeExpressionNested(attr.Expr, ctx, fieldV.Addr().Interface())...)
+			diags = append(diags, DecodeExpressionNested(
+				attr.Expr, ctx, fieldV.Addr().Interface())...)
 		}
 	}
 
@@ -278,74 +270,36 @@ func decodeBlockToValue(block *hcl.Block, ctx *hcl.EvalContext, v reflect.Value)
 	return diags
 }
 
-func DecodeExpressionNested(expr hcl.Expression, ctx *hcl.EvalContext, val interface{}) hcl.Diagnostics {
-	defer func() {
-		if r := recover(); r != nil {
-			fmt.Println(r)
-		}
-	}()
-
-	var diags hcl.Diagnostics
-
-	//var shouldDecode = true
-	//if isStruct(val) {
-	//	_, shouldDecode = val.(modconfig.CtyValueProvider)
-	//}
-	//if shouldDecode {
-	// TODO use gohcl fxn
-	diags = decodeExpression(expr, ctx, val)
-	if diags.HasErrors() {
-		return diags
-	}
-	//}
-
-	nested := getNestedStructVals(val)
-	for _, v := range nested {
-		moreDiags := DecodeExpressionNested(expr, ctx, v)
-		diags = append(diags, moreDiags...)
-	}
-	return diags
-
-}
-
-//func containsTag(val interface{}, tagName string) (bool, reflect.Kind) {
-//	// check the struct has the given tag
-//	rv := reflect.ValueOf(val)
-//	//rt := rv.Type()
-//	kind := rv.Kind()
-//	// dereference pointers
-//	for kind == reflect.Pointer {
-//		rv = rv.Elem()
-//		kind = rv.Kind()
-//		log.Printf("[WARN] containsTag kind %s", rv.Kind())
-//	}
+// DecodeExpression extracts the value of the given expression into the given
+// value. This value must be something that gocty is able to decode into,
+// since the final decoding is delegated to that package.
 //
-//	if kind != reflect.Struct {
-//		return false, kind
-//	}
-//	for i := 0; i < rv.Type().NumField(); i++ {
-//		field := rv.Type().Field(i)
-//		if tag := field.Tag.Get(tagName); tag != "" {
-//			return true, kind
-//		}
-//	}
+// The given EvalContext is used to resolve any variables or functions in
+// expressions encountered while decoding. This may be nil to require only
+// constant values, for simple applications that do not support variables or
+// functions.
 //
-//	return false, kind
-//}
+// The returned diagnostics should be inspected with its HasErrors method to
+// determine if the populated value is valid and complete. If error diagnostics
+// are returned then the given value may have been partially-populated but
+// may still be accessed by a careful caller for static analysis and editor
+// integration use-cases.
+//
+// NOTE: change in behaviour from gohcl - if no ImpliedType can be determined, do not panic, just retuirn nil
+// this is to handle the case where we are trying to decode into a nested struct type which has no cty tags
+func DecodeExpression(expr hcl.Expression, ctx *hcl.EvalContext, val interface{}) hcl.Diagnostics {
+	srcVal, diags := expr.Value(ctx)
 
-func decodeExpression(expr hcl.Expression, ctx *hcl.EvalContext, val interface{}) hcl.Diagnostics {
-	var diags hcl.Diagnostics
-	exprVal, diags := expr.Value(ctx)
-	if diags.HasErrors() {
-		return diags
-	}
 	convTy, err := gocty.ImpliedType(val)
 	if err != nil {
+		// NOTE: gohcl panics here
 		// we may hit this if we try to decode into a base struct with no cty tags
+		// so just return nil
 		return nil
 		//panic(fmt.Sprintf("unsuitable DecodeExpression target: %s", err))
 	}
-	exprVal, err = convert.Convert(exprVal, convTy)
+
+	srcVal, err = convert.Convert(srcVal, convTy)
 	if err != nil {
 		diags = append(diags, &hcl.Diagnostic{
 			Severity: hcl.DiagError,
@@ -357,7 +311,7 @@ func decodeExpression(expr hcl.Expression, ctx *hcl.EvalContext, val interface{}
 		return diags
 	}
 
-	err = gocty.FromCtyValue(exprVal, val)
+	err = gocty.FromCtyValue(srcVal, val)
 	if err != nil {
 		diags = append(diags, &hcl.Diagnostic{
 			Severity: hcl.DiagError,
@@ -369,175 +323,4 @@ func decodeExpression(expr hcl.Expression, ctx *hcl.EvalContext, val interface{}
 	}
 
 	return diags
-}
-
-// ImpliedBodySchema produces a hcl.BodySchema derived from the type of the
-// given value, which must be a struct value or a pointer to one. If an
-// inappropriate value is passed, this function will panic.
-//
-// The second return argument indicates whether the given struct includes
-// a "remain" field, and thus the returned schema is non-exhaustive.
-//
-// This uses the tags on the fields of the struct to discover how each
-// field's value should be expressed within configuration. If an invalid
-// mapping is attempted, this function will panic.
-func ImpliedBodySchema(val interface{}) (schema *hcl.BodySchema, partial bool) {
-	ty := reflect.TypeOf(val)
-
-	if ty.Kind() == reflect.Ptr {
-		ty = ty.Elem()
-	}
-
-	if ty.Kind() != reflect.Struct {
-		panic(fmt.Sprintf("given value must be struct, not %T", val))
-	}
-
-	var attrSchemas []hcl.AttributeSchema
-	var blockSchemas []hcl.BlockHeaderSchema
-
-	tags := getFieldTags(ty)
-
-	attrNames := make([]string, 0, len(tags.Attributes))
-	for n := range tags.Attributes {
-		attrNames = append(attrNames, n)
-	}
-	sort.Strings(attrNames)
-	for _, n := range attrNames {
-		idx := tags.Attributes[n]
-		optional := tags.Optional[n]
-		field := ty.Field(idx)
-
-		var required bool
-
-		switch {
-		case field.Type.AssignableTo(exprType):
-			// If we're decoding to hcl.Expression then absense can be
-			// indicated via a null value, so we don't specify that
-			// the field is required during decoding.
-			required = false
-		case field.Type.Kind() != reflect.Ptr && !optional:
-			required = true
-		default:
-			required = false
-		}
-
-		attrSchemas = append(attrSchemas, hcl.AttributeSchema{
-			Name:     n,
-			Required: required,
-		})
-	}
-
-	blockNames := make([]string, 0, len(tags.Blocks))
-	for n := range tags.Blocks {
-		blockNames = append(blockNames, n)
-	}
-	sort.Strings(blockNames)
-	for _, n := range blockNames {
-		idx := tags.Blocks[n]
-		field := ty.Field(idx)
-		fty := field.Type
-		if fty.Kind() == reflect.Slice {
-			fty = fty.Elem()
-		}
-		if fty.Kind() == reflect.Ptr {
-			fty = fty.Elem()
-		}
-		if fty.Kind() != reflect.Struct {
-			panic(fmt.Sprintf(
-				"hcl 'block' tag kind cannot be applied to %s field %s: struct required", field.Type.String(), field.Name,
-			))
-		}
-		ftags := getFieldTags(fty)
-		var labelNames []string
-		if len(ftags.Labels) > 0 {
-			labelNames = make([]string, len(ftags.Labels))
-			for i, l := range ftags.Labels {
-				labelNames[i] = l.Name
-			}
-		}
-
-		blockSchemas = append(blockSchemas, hcl.BlockHeaderSchema{
-			Type:       n,
-			LabelNames: labelNames,
-		})
-	}
-
-	partial = tags.Remain != nil
-	schema = &hcl.BodySchema{
-		Attributes: attrSchemas,
-		Blocks:     blockSchemas,
-	}
-	return schema, partial
-}
-
-type fieldTags struct {
-	Attributes map[string]int
-	Blocks     map[string]int
-	Labels     []labelField
-	Remain     *int
-	Body       *int
-	Optional   map[string]bool
-}
-
-type labelField struct {
-	FieldIndex int
-	Name       string
-}
-
-func getFieldTags(ty reflect.Type) *fieldTags {
-	ret := &fieldTags{
-		Attributes: map[string]int{},
-		Blocks:     map[string]int{},
-		Optional:   map[string]bool{},
-	}
-
-	ct := ty.NumField()
-	for i := 0; i < ct; i++ {
-		field := ty.Field(i)
-		tag := field.Tag.Get("hcl")
-		if tag == "" {
-			continue
-		}
-
-		comma := strings.Index(tag, ",")
-		var name, kind string
-		if comma != -1 {
-			name = tag[:comma]
-			kind = tag[comma+1:]
-		} else {
-			name = tag
-			kind = "attr"
-		}
-
-		switch kind {
-		case "attr":
-			ret.Attributes[name] = i
-		case "block":
-			ret.Blocks[name] = i
-		case "label":
-			ret.Labels = append(ret.Labels, labelField{
-				FieldIndex: i,
-				Name:       name,
-			})
-		case "remain":
-			if ret.Remain != nil {
-				panic("only one 'remain' tag is permitted")
-			}
-			idx := i // copy, because this loop will continue assigning to i
-			ret.Remain = &idx
-		case "body":
-			if ret.Body != nil {
-				panic("only one 'body' tag is permitted")
-			}
-			idx := i // copy, because this loop will continue assigning to i
-			ret.Body = &idx
-		case "optional":
-			ret.Attributes[name] = i
-			ret.Optional[name] = true
-		default:
-			panic(fmt.Sprintf("invalid hcl field tag kind %q on %s %q", kind, field.Type.String(), field.Name))
-		}
-	}
-
-	return ret
 }
