+++ pkg/steampipeconfig/parse/parse_hcl/types.go
@@ -1,0 +1,15 @@
+package parse_hcl
+
+import (
+	"github.com/hashicorp/hcl/v2"
+	"reflect"
+)
+
+var victimExpr hcl.Expression
+var victimBody hcl.Body
+
+var exprType = reflect.TypeOf(&victimExpr).Elem()
+var bodyType = reflect.TypeOf(&victimBody).Elem()
+var blockType = reflect.TypeOf((*hcl.Block)(nil))
+var attrType = reflect.TypeOf((*hcl.Attribute)(nil))
+var attrsType = reflect.TypeOf(hcl.Attributes(nil))
